+++ drivers.c
@@ -1115,7 +1115,7 @@ static bool mtk3301_rate_switcher(struct gps_device_t *session, double rate)
 {
     char buf[78];
 
-    unsigned int milliseconds = 1000 * rate;
+    unsigned int milliseconds = (unsigned int)(1000 * rate);
     if (rate > 1)
 	milliseconds = 1000;
     else if (rate < 0.2)
